Fix the change comparison during change merge.
Before this fix, the --merge-change task compared the changed entities
by type, id and version. The result was that merging two change
streams with the same entities but different version resulted in
_both_ changes occuring on the resultant entity stream, which is
pretty useless for a merge (for example, the version conflict
resolution methods never produced any sensible result). Fix this by
comparing by just type and id and hand off version handling to the
merger code already in place (much like in the entity merger).